requirements.txt
google-cloud-storage==2.7.0
google-auth==2.16.3
google-resumable-media==2.4.1
google-api-core==2.11.0
cachetools==5.2.0
google-cloud==0.34.0
google-cloud-core==2.3.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
six==1.16.0
rsa==4.8
googleapis-common-protos==1.56.3
google-crc32c==1.3.0
protobuf==4.22.1

Flask==2.2.3
importlib-metadata==4.12.0
zipp==1.0.0
more-itertools==8.10.0
requests==2.28.2
Jinja2==3.1.2
itsdangerous==2.1.2
Werkzeug==2.2.3
click==8.1.3
MarkupSafe==2.1.2
urllib3==1.26.15
certifi==2022.12.7
idna==3.4
pipreqs==0.4.11
yarg==0.1.9
charset-normalizer==3.1.0
docopt==0.6.2